Fix bug in limit comparison
The previous code tried to remove entries while the limit was greater than the count. The intended operation is to remove entries while the count is greater than the limit. The zero is used as unlimited. An added bonus of getting the comparison direction correct, is that it can be more succinctly reformatted as a chained comparison.