begin transition from built in feedgen
The Feedgen generator rss was not being validated by podcast catchers; after much testing, it's entirely possible that it is likely due to not assigning pubdates to the items in the feed, but it was also missing an atom self-link and some itunes tags. It didn't seem to have a way to support subcategories correctly for itunes either. So anyway, basically I'm planning on transitioning to a more bare-bones object representation that just uses string formatting and many properties to produce the rss feed.